renamed reader classes
